[#14] Optimize Home's header & footer content, remove HomeSurveys
