+++ android/src/main/java/vn/luongvo/kmm/survey/android/ui/screens/home/SurveyUiModel.kt
@@ -1,0 +1,7 @@
+package vn.luongvo.kmm.survey.android.ui.screens.home
+
+data class SurveyUiModel(
+    val title: String,
+    val description: String,
+    val imageUrl: String
+)
